+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.42.0 (Unreleased)
+FEATURES:
+* mdb: support Redis 6.0 in `yandex_mdb_redis_cluster`
 
 ENHANCEMENTS:
 * mdb: add `role` attribute to `host` entity in `yandex_mdb_postgresql_cluster` resource and data source